aligner: add small utility to calculate offsets
aligner introduces a small function that takes in the number of frames,
and desired window-size and produces two starting indexes that are misaligned
(not equal). This is done randomly.
The starting indexes guarantee that a window_size of samples is still available
with that starting index.
There is an optional paramter that regulates whether the resulting windows can
overlap or not.
Lastly, it adds a small unittest to test both of those behaviors.